Source Code/Frontend/requirements.txt
Flask
flask-cors
tensorflow
numpy
ultralytics 
Pillow


